logs zoom change
