+++ apps/src/codebridge/Codebridge.tsx
@@ -14,6 +14,7 @@ import {
   CodebridgeLevelProperties,
 } from '@codebridge/types';
 import classNames from 'classnames';
+import debounce from 'lodash/debounce';
 import React, {useEffect, useMemo, useReducer, useRef} from 'react';
 
 import {LabConfig, MultiFileSource, ProjectSources} from '@cdo/apps/lab2/types';
@@ -23,6 +24,8 @@ import {useAppSelector} from '@cdo/apps/util/reduxHooks';
 
 import moduleStyles from './styles/codebridgeContainer.module.scss';
 import './styles/codebridge.scss';
+
+const DEBOUNCE_TIME_OUT = 300;
 
 type CodebridgeProps = {
   source: MultiFileSource;
@@ -87,6 +90,57 @@ export const Codebridge = React.memo(
       () => new BackpackClientApi(appName, null),
       [appName]
     );
+    const getInitialDevicePixelRatio = (): number => {
+      return window?.devicePixelRatio || 1;
+    };
+
+    const detectZoom = (initialDPR: number): number[] => {
+      const currentDPR = window.devicePixelRatio || 1;
+      const zoomDPR = currentDPR / initialDPR;
+      const zoomValues = [Math.round(zoomDPR * 100), 100];
+      if (window.visualViewport?.scale) {
+        zoomValues[1] = Math.round(window.visualViewport.scale * 100);
+      }
+      return zoomValues;
+    };
+
+    useEffect(() => {
+      const initialDPR = getInitialDevicePixelRatio();
+      let lastZoomValues = detectZoom(initialDPR);
+
+      const logZoomChange = (zoomPercent: number, direction: 'in' | 'out') => {
+        console.log('BrowserZoomChanged', {
+          zoomPercent,
+          direction,
+          appName,
+        });
+      };
+
+      const checkZoom = () => {
+        const currentZoomValues = detectZoom(initialDPR);
+        let loggedZoom = false;
+        currentZoomValues.forEach((zoomValue, index) => {
+          if (zoomValue !== lastZoomValues[index] && !loggedZoom) {
+            const direction =
+              currentZoomValues[0] > lastZoomValues[0] ? 'in' : 'out';
+            logZoomChange(currentZoomValues[index], direction);
+            lastZoomValues = currentZoomValues;
+            loggedZoom = true;
+          }
+        });
+      };
+      const debouncedCheckZoom = debounce(checkZoom, DEBOUNCE_TIME_OUT);
+      window.visualViewport?.addEventListener('resize', debouncedCheckZoom);
+
+      return () => {
+        debouncedCheckZoom.cancel();
+        // clearInterval(interval);
+        window.visualViewport?.removeEventListener(
+          'resize',
+          debouncedCheckZoom
+        );
+      };
+    }, [appName]);
 
     return (
       <CodebridgeContextProvider
